Fix return value of `T::Private:Methods.signature_for_method`
Since Sorbet runtime and all signatures are erased, this method should return `nil` to signal that there is no signature for the given method.

Returning `method` here is not correct, since the return value is expected to be a `Signature` instance that has a reference to the original method in the `method` property. An `UnboundMethod` instance does not satisfy this property.