`T::Props#props` is expected to be a hash not an array
`T::Props#props` is public API of `T::Struct` and is used to enumerate all the properties defined on a typed struct.

The original implementation of this method, however, returns a hash with method name keys and rules as the value. Returning an array as the value of `props` violates this expectation.